requirements.txt
# SENTINEL v5.0 - Complete Requirements
# Tested and verified on NVIDIA GTX 1650

# Core
numpy==1.24.3
pillow==10.1.0
opencv-python==4.8.1.78

# Deep Learning (Install separately with CUDA)
# torch==2.1.0+cu118
# torchvision==0.16.0+cu118
ultralytics==8.1.0

# Face Recognition
insightface==0.7.3
onnxruntime-gpu==1.16.3

# Tracking
filterpy==1.4.5
scipy==1.11.4
lap==0.4.0

# Pose Estimation
mediapipe==0.10.8

# Audio
pyttsx3==2.90

# Configuration
omegaconf==2.3.0

# Logging
loguru==0.7.2

# Utilities
psutil==5.9.6
